Moved test utils into their own directory
